requirements.txt
blinker==1.7.0
click==8.1.7
dnspython==2.4.2
flask==3.0.0
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonify==0.5
MarkupSafe==2.1.3
pymongo==4.6.0
python-dotenv==1.0.0
werkzeug==3.0.1
zipp==3.17.0
